requirements.txt
gitpython >= 3.1.8
munch >= 2.5.0
pyyaml >= 5.3.1